Make joint feedback checker header only
